requirement.txt
requests~=2.28.1
python-dotenv~=0.20.0
asyncio~=3.4.3
python-telegram-bot~=20.0a2